refactor: separate FileFlags from cli.go
